20240627test4/requirements.txt
albumentations==1.4.11
comet_ml==3.44.1
coremltools==7.2
Flask==3.0.3
GitPython==3.1.43
ipython==8.12.3
joblib==1.2.0
matplotlib==3.7.1
mss==9.0.1
nncf==2.11.0
numpy==1.25.0
onnx==1.16.1
onnxruntime==1.18.1
onnxsim==0.4.36
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
openvino==2024.2.0
pafy==0.5.5
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.0
PyQt5==5.15.10
PyQt5_sip==12.11.0
pywin32==305
PyYAML==6.0
Requests==2.32.3
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
setuptools==60.2.0
tensorflow==2.17.0
tensorflowjs==4.20.0
tensorrt==10.2.0.post1
tflite_runtime==2.14.0
tflite_support==0.4.4
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.2
train==0.0.5
tritonclient==2.47.0
ultralytics==8.2.48
ultralytics_thop==2.0.0
x2paddle==1.4.1
